+++ src/openms/include/OpenMS/ML/GRIDSEARCH/sources.cmake
@@ -1,16 +1,9 @@
 ### the directory name
-set(directory include/OpenMS/MATH/MISC)
+set(directory include/OpenMS/ML/GRIDSEARCH)
 
 ### list all header files of the directory here
 set(sources_list_h
-BilinearInterpolation.h
-BSpline2d.h
-CubicSpline2d.h
-EmgGradientDescent.h
-LinearInterpolation.h
-MathFunctions.h
-MSNumpress.h
-SplineBisection.h
+GridSearch.h
 )
 
 ### add path to the filenames
@@ -20,6 +13,6 @@ foreach(i ${sources_list_h})
 endforeach(i)
 
 ### source group definition
-source_group("Header Files\\OpenMS\\MATH\\MISC" FILES ${sources_h})
+source_group("Header Files\\OpenMS\\ML\\GRIDSEARCH" FILES ${sources_h})
 
 set(OpenMS_sources_h ${OpenMS_sources_h} ${sources_h})
