+++ src/openms/includes.cmake
@@ -10,6 +10,7 @@ include(source/ML/sources.cmake)
 include(source/ML/NNLS/sources.cmake)
 include(source/ML/SVM/sources.cmake)
 include(source/ML/CLUSTERING/sources.cmake)
+include(source/ML/GRIDSEARCH/sources.cmake)
 include(source/ML/RANSAC/sources.cmake)
 include(source/DATASTRUCTURES/sources.cmake)
 include(source/METADATA/sources.cmake)
@@ -89,6 +90,7 @@ include(include/OpenMS/ANALYSIS/TOPDOWN/sources.cmake)
 include(include/OpenMS/ANALYSIS/XLMS/sources.cmake)
 
 include(include/OpenMS/ML/CLUSTERING/sources.cmake)
+include(include/OpenMS/ML/GRIDSEARCH/sources.cmake)
 include(include/OpenMS/COMPARISON/SPECTRA/sources.cmake)
 include(include/OpenMS/FILTERING/BASELINE/sources.cmake)
 include(include/OpenMS/FILTERING/CALIBRATION/sources.cmake)
