+++ src/openms/source/ML/SVM/SimpleSVM.cpp
@@ -13,7 +13,7 @@
 #include <OpenMS/DATASTRUCTURES/ListUtils.h>
 
 #include <OpenMS/MATH/STATISTICS/StatisticFunctions.h>
-#include <OpenMS/MATH/MISC/GridSearch.h>
+#include <OpenMS/ML/GRIDSEARCH/GridSearch.h>
 
 #include <cstdlib>
